requirements.txt
wheel
flask
gunicorn
mysqlclient
flask-sqlalchemy
flask-login
flask-mysqldb
